requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-phonenumber-field==6.1.0     
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
phonenumbers==8.12.48
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
PyJWT==1.7.1
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
